Shuffle around loads of stuff
Just to tidy things up. Expect more on this branch.
