+++ config.ru
@@ -1,3 +1,3 @@
-require_relative 'lib/sir_handel/app'
+require_relative 'lib/sir_handel'
 
 run SirHandel::App
