+++ spec/spec_helper.rb
@@ -1,13 +1,22 @@
 ENV['RACK_ENV'] = 'test'
-
-require_relative 'support/vcr_setup'
-require 'coveralls'
-Coveralls.wear_merged!
-
 # Comment out when recording new VCR cassettes
 ENV['ES_URL'] = 'http://elastic.search/'
 
+require 'rack/test'
+require 'coveralls'
+Coveralls.wear_merged!
+require_relative 'support/vcr_setup'
+
+require File.expand_path '../../lib/sir_handel.rb', __FILE__
+
+module RSpecMixin
+  include Rack::Test::Methods
+  def app() SirHandel::App end
+end
+
 RSpec.configure do |config|
+  config.include RSpecMixin
+
   config.after :each do
     Blocktrain::Lookups.instance.reset!
   end
